Some annotations refactoring
