+++ src/main/java/com/googlecode/jsonrpc4j/AnnotationsErrorResolver.java
@@ -24,6 +24,8 @@ THE SOFTWARE.
 
 package com.googlecode.jsonrpc4j;
 
+import com.googlecode.jsonrpc4j.annotations.JsonRpcErrors;
+import com.googlecode.jsonrpc4j.annotations.JsonRpcError;
 import java.lang.reflect.Method;
 import java.util.List;
