+++ src/main/java/com/googlecode/jsonrpc4j/JsonRpcHttpClient.java
@@ -110,6 +110,7 @@ public class JsonRpcHttpClient
 	 * @param arguments the arguments to the method
 	 * @throws Throwable on error
 	 */
+    @Override
 	public void invoke(String methodName, Object argument)
 		throws Throwable {
 		invoke(methodName, argument, null, new HashMap<String, String>());
@@ -126,6 +127,7 @@ public class JsonRpcHttpClient
 	 * @return the return value
 	 * @throws Throwable on error
 	 */
+    @Override
 	public Object invoke(
 		String methodName, Object argument, Type returnType)
 		throws Throwable {
@@ -144,6 +146,7 @@ public class JsonRpcHttpClient
 	 * @throws Throwable on error
 	 */
 	@SuppressWarnings("unchecked")
+    @Override
 	public <T> T invoke(
 		String methodName, Object argument, Class<T> clazz)
 		throws Throwable {
@@ -162,6 +165,7 @@ public class JsonRpcHttpClient
 	 * @return the return value
 	 * @throws Throwable on error
 	 */
+    @Override
 	public Object invoke(
 		String methodName, Object argument, Type returnType,
 		Map<String, String> extraHeaders)
@@ -227,6 +231,7 @@ public class JsonRpcHttpClient
 	 * @throws Throwable on error
 	 */
 	@SuppressWarnings("unchecked")
+    @Override
 	public <T> T invoke(
 		String methodName, Object argument, Class<T> clazz,
 		Map<String, String> extraHeaders)
