+++ src/main/java/com/googlecode/jsonrpc4j/spring/AutoJsonRpcClientProxyCreator.java
@@ -25,7 +25,7 @@ THE SOFTWARE.
 package com.googlecode.jsonrpc4j.spring;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.googlecode.jsonrpc4j.JsonRpcService;
+import com.googlecode.jsonrpc4j.annotations.JsonRpcService;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.springframework.beans.BeansException;
