+++ src/test/java/com/googlecode/jsonrpc4j/JsonRpcServerTest.java
@@ -1,5 +1,8 @@
 package com.googlecode.jsonrpc4j;
 
+import com.googlecode.jsonrpc4j.annotations.JsonRpcMethod;
+import com.googlecode.jsonrpc4j.annotations.JsonRpcParamName;
+import com.googlecode.jsonrpc4j.annotations.JsonRpcParam;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.TextNode;
