+++ src/test/java/com/googlecode/jsonrpc4j/MultiServiceTest.java
@@ -1,5 +1,6 @@
 package com.googlecode.jsonrpc4j;
 
+import com.googlecode.jsonrpc4j.annotations.JsonRpcParam;
 import static org.junit.Assert.assertEquals;
 
 import java.io.ByteArrayOutputStream;
