+++ src/test/java/com/googlecode/jsonrpc4j/spring/Service.java
@@ -1,6 +1,6 @@
 package com.googlecode.jsonrpc4j.spring;
 
-import com.googlecode.jsonrpc4j.JsonRpcService;
+import com.googlecode.jsonrpc4j.annotations.JsonRpcService;
 
 @JsonRpcService("Service")
 interface Service {
